moved actions and cleaned up
