+++ packages/i18n/locales/en/translation.json
@@ -32,6 +32,8 @@
     "rebalancerLabel": "Rebalancer",
     "smartContractingDescription": "Create, execute, and manage smart contracts.",
     "smartContractingLabel": "Smart Contracting",
+    "subDaosDescription": "Manage SubDAOs or become a SubDAO.",
+    "subDaosLabel": "SubDAOs",
     "treasuryDescription_dao": "Manage this DAO's treasury (for example, send payments, stake or swap tokens, etc.).",
     "treasuryDescription_gov": "Manage the community pool (for example, send payments, stake or swap tokens, etc.).",
     "treasuryDescription_wallet": "Manage your tokens (for example, send payments, stake or swap tokens, etc.).",
@@ -315,6 +317,7 @@
     "cameraWithFlash": "Camera with flash",
     "chains": "Chains",
     "chart": "Chart",
+    "check": "Check",
     "clock": "Clock",
     "closedLockWithKey": "Closed lock with key",
     "computerDisk": "Computer disk",
@@ -537,7 +540,7 @@
     "unsupportedWallet": "Unsupported wallet. Please use a different wallet."
   },
   "form": {
-    "acceptSubDaoAddressInputLabel": "SubDAO to approve",
+    "acceptSubDaoAddressInputLabel": "SubDAO to accept",
     "addAnImage": "Add an image",
     "address": "Address",
     "addressInputPlaceholder_any": "Search profile or DAO name or enter an address...",
@@ -568,7 +571,7 @@
     "automaticallyAddTokensTitle": "Automatically add tokens",
     "automaticallyAddTokensTooltip": "Should tokens sent to the DAO get added to the treasury?",
     "baseToken": "Base token",
-    "becomeSubDaoAdminInputLabel": "DAO to join",
+    "becomeSubDaoAdminInputLabel": "New parent DAO",
     "blocksToPauseFor": "Blocks to pause for",
     "buttonLabel": "Button label",
     "calls": "Calls",
@@ -897,8 +900,8 @@
   },
   "info": {
     "abstainVote": "Abstain",
-    "acceptSubDaoActionDescription": "Approve the ownership transfer request from your SubDAO. Be sure that your SubDAO has already requested to join your DAO through the \"Become SubDAO\" action.\"",
-    "acceptSubDaoDescription": "Approve SubDAO ownership transfer.",
+    "acceptSubDaoActionDescription": "Accept a pending SubDAO's ownership transfer request. Be sure that the SubDAO-to-be has already requested to join your DAO through the \"Become SubDAO\" action.",
+    "acceptSubDaoDescription": "Accept a DAO's request to become a SubDAO.",
     "acknowledgeServiceFee": "We acknowledge that {{fee}} will be charged from the treasury for registration to the rebalancer.",
     "acknowledgeServiceFee_none": "We acknowledge that there is no fee to register for the rebalancer.",
     "actionLibraryDescription": "Choose one or more actions to perform upon execution.",
@@ -948,8 +951,8 @@
     "available": "available",
     "availableBalance": "Available balance",
     "availableHistoryLoaded": "All available history has been loaded.",
-    "becomeSubDaoActionDescription": "Provide your parent DAO address to request join as its SubDAO. Your parent DAO will need to approve your request to complete the ownership transfer.",
-    "becomeSubDaoDescription": "Request to join DAO as SubDAO.",
+    "becomeSubDaoActionDescription": "Provide the address of a DAO to request to become its SubDAO. The DAO will then need to accept your request to complete the ownership transfer and become your parent. Once the transfer is complete, only your new parent DAO can transfer ownership in the future. Keep in mind that your parent DAO will have full authority to execute anything it wants to on behalf of this DAO.",
+    "becomeSubDaoDescription": "Request to become a SubDAO of a DAO.",
     "betterNonPfmIbcPathAvailable": "A better path exists that unwinds the IBC token properly, but it requires hops that cannot be performed in a single proposal due to the configuration of one or more of the chains involved. You may safely use the direct path above, though it likely results in an undesirable IBC token on the final chain that will eventually need to be unwound. If you wish to unwind it manually, follow the path below by creating an individual proposal for the first hop, then a second proposal once the first one is passed and executed, and so on.",
     "betterPfmIbcPathAvailable_one": "A better path exists that unwinds the IBC token properly, but it requires you to have an account on {{chains}}. An account must exist on each intermediary chain in a multi-hop transaction to act as a failsafe in the event of catastrophic failure (unlikely). Either a cross-chain or ICA account works and will be automatically detected once created. You may safely use the direct path above, though it likely results in an undesirable IBC token on the final chain that will eventually need to be unwound.",
     "betterPfmIbcPathAvailable_other": "A better path exists that unwinds the IBC token properly, but it requires you to have accounts on the following chains: {{chains}}. An account must exist on each intermediary chain in a multi-hop transaction to act as a failsafe in the event of catastrophic failure (unlikely). Either a cross-chain or ICA account works and will be automatically detected once created. You may safely use the direct path above, though it likely results in an undesirable IBC token on the final chain that will eventually need to be unwound.",
@@ -1570,7 +1573,7 @@
   "title": {
     "404": "404: Not Found",
     "500": "500: Error",
-    "acceptSubDao": "Approve SubDAO join",
+    "acceptSubDao": "Accept SubDAO Request",
     "accepted": "Accepted",
     "acceptingSubmissions": "Accepting submissions",
     "account": "Account",
