+++ packages/stateful/actions/core/subdaos/AcceptSubDao/Component.stories.tsx
@@ -3,17 +3,17 @@ import { ComponentMeta, ComponentStory } from '@storybook/react'
 import { ReactHookFormDecorator } from '@dao-dao/storybook'
 
 import { AddressInput } from '../../../../components'
-import { BecomeSubDaoComponent } from './Component'
+import { AcceptSubDaoComponent } from './Component'
 
 export default {
   title:
-    'DAO DAO / packages / stateful / actions / core / dao_governance / BecomeSubDao',
-  component: BecomeSubDaoComponent,
+    'DAO DAO / packages / stateful / actions / core / subdaos / AcceptSubDao',
+  component: AcceptSubDaoComponent,
   decorators: [ReactHookFormDecorator],
-} as ComponentMeta<typeof BecomeSubDaoComponent>
+} as ComponentMeta<typeof AcceptSubDaoComponent>
 
-const Template: ComponentStory<typeof BecomeSubDaoComponent> = (args) => (
-  <BecomeSubDaoComponent {...args} />
+const Template: ComponentStory<typeof AcceptSubDaoComponent> = (args) => (
+  <AcceptSubDaoComponent {...args} />
 )
 
 export const Default = Template.bind({})
