+++ packages/stateful/actions/core/subdaos/ManageSubDaos/Component.stories.tsx
@@ -8,7 +8,7 @@ import { ManageSubDaosComponent } from './Component'
 
 export default {
   title:
-    'DAO DAO / packages / stateful / actions / core / dao_governance / ManageSubDaos',
+    'DAO DAO / packages / stateful / actions / core / subdaos / ManageSubDaos',
   component: ManageSubDaosComponent,
   decorators: [ReactHookFormDecorator],
 } as ComponentMeta<typeof ManageSubDaosComponent>
